+++ client/js/components/Header.js
@@ -1,13 +1,14 @@
 const {Nav, NavItem, NavDropdown, MenuItem, Navbar} = ReactBootstrap
 import NetworkStatus from './NetworkStatus'
+import Basic from './Basic'
 
-class Header extends React.Component {
+class Header extends Basic {
 
   render() {
 
     let dropDown
 
-    const ps = this.props.appState
+    const ps = this.appState()
     const wallet = ps.wallet
     let twitter
     try {
@@ -17,7 +18,7 @@ class Header extends React.Component {
 
     if (wallet) {
 
-      if (twitter) {
+      if (twitter && twitter.name) {
         dropDown = <NavDropdown eventKey={3} title={<img src={twitter.avatar} className="tavatar circled"/>}
                                 id="basic-nav-dropdown">
           <li role="presentation">
@@ -30,7 +31,7 @@ class Header extends React.Component {
       } else {
         dropDown = <Nav>
           <NavItem eventKey={1} href="#">
-            {'Wallet ' + this.props.appState.wallet.substring(0, 6) + ' (anonymous)'}
+            {'Wallet ' + this.appState().wallet.substring(0, 6) + ' (anonymous)'}
           </NavItem>
         </Nav>
       }
@@ -38,7 +39,7 @@ class Header extends React.Component {
 
     return (
       <div>
-        <NetworkStatus appState={this.props.appState}/>
+        <NetworkStatus appState={this.props.app.appState}/>
         <Navbar
           staticTop
           componentClass="header"
