+++ client/js/components/extras/LoadingBadge.js
