+++ static/debug-index.html
@@ -4,11 +4,9 @@
 
   <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
 
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/antd/3.5.2/antd.css" crossorigin="anonymous">
   <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
         crossorigin="anonymous">
 
-  <link rel="stylesheet" href="//fonts.googleapis.com/css?family=Lato"/>
   <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css"/>
   <link rel="stylesheet" href="/css/style.css"/>
   <link rel="icon" href="favicon.ico?v=2"/>
